requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#

numpy==1.17.4             # via opencv-python
opencv-python==4.1.2.30
pillow==6.2.1
torch==1.0.1
